Add search implementation
