+++ appJs/appApiJs/build.gradle.kts
@@ -2,7 +2,7 @@ import com.gchristov.thecodinglove.gradleplugins.Deps
 
 plugins {
     id("javascript-node-executable-plugin")
-    // TODO: This shouldn't be needed here
+    // TODO: This shouldn't be needed here but is used for the Firestore example
     id("kmp-data-plugin")
 }
 
@@ -12,7 +12,7 @@ kotlin {
             dependencies {
                 // TODO: Use feature modules here when ready
                 implementation(projects.kmpCommonFirebase)
-                implementation(projects.kmpSearchData)
+                implementation(projects.kmpSearch)
                 // Ideally these would be linked from corresponding submodules but that is currently
                 // not supported out of the box or through the npm-publish plugin and causes "module
                 // not found" errors. As a workaround, all NPM dependencies will be listed here,
