+++ appJs/appApiJs/src/jsMain/kotlin/com/gchristov/thecodinglove/JavascriptMainApi.kt
@@ -1,10 +1,8 @@
 package com.gchristov.thecodinglove
 
 import com.gchristov.thecodinglove.kmpcommonfirebase.CommonFirebaseModule
-import com.gchristov.thecodinglove.kmpcommonfirebase.CommonNetworkModule
-import com.gchristov.thecodinglove.kmpsearchdata.SearchDataModule
-import io.ktor.client.call.*
-import io.ktor.client.request.*
+import com.gchristov.thecodinglove.kmpsearch.SearchHistory
+import com.gchristov.thecodinglove.kmpsearch.SearchModule
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import kotlinx.serialization.Serializable
@@ -12,11 +10,9 @@ import kotlinx.serialization.Serializable
 internal actual fun serveApi(args: Array<String>) {
     val fireFunctions = require("firebase-functions")
     exports.myTestFun = fireFunctions.https.onRequest { request, response ->
-        val client = CommonNetworkModule.injectHttpClient()
         // TODO: Do not use GlobalScope
+        val searchQuery = request.query.searchQuery as? String
         GlobalScope.launch {
-            val userResponse: Response = client.get("https://reqres.in/api/users").body()
-
             println("About to test Firestore")
             val firestore = CommonFirebaseModule.injectFirestore()
             val document = firestore.document("preferences/user1").get()
@@ -26,558 +22,23 @@ internal actual fun serveApi(args: Array<String>) {
             batch.set(firestore.document("preferences/user1"), Count(count + 1))
             batch.commit()
 
-            println("About to test html parser")
-            val parser = SearchDataModule.injectPostParser()
-            val totalCount = parser.parseTotalPosts(TestHtml)
-            val results = parser.parsePosts(TestHtml)
-            val parseResult = " parseCount=$totalCount parsePosts=$results"
+            println("About to test search")
+            val search = SearchModule.injectSearchUseCase()
+            val searchResult = search(
+                query = searchQuery ?: "release",
+                searchHistory = SearchHistory(),
+                resultsPerPage = 4
+            )
 
-            response.send("{networkResponsePage=${userResponse.page}, invocationsCount=$count, $parseResult}")
+            response.send("{\"invocationsCount\": $count, \"searchResult\": \"$searchResult\"}")
         }
     }
 }
 
-private external fun require(module:String) : dynamic
+private external fun require(module: String): dynamic
 private external var exports: dynamic
-
-@Serializable
-private data class Response(
-    val page: Int
-)
 
 @Serializable
 private data class Count(
     val count: Int
 )
-
-private val TestHtml = """
-    <!doctype html><html
-    lang=en><head><style>img.lazy{min-height:1px}</style><link
-    rel=preload href=https://thecodinglove.com/wp-content/plugins/w3-total-cache/pub/js/lazyload.min.js as=script><meta
-    charset="utf-8"><meta
-    name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no"><meta
-    name="description" content="The Coding Love - Programmer humor: gifs, memes, jokes"><meta
-    name="keywords" content="programmer,developer,humor,gif,meme,joke,thecodinglove,coding,programming,computer"><title>
-    When the sales guy is about to click on a button I never tested | The Coding Love - Programmer humor: gifs, memes, jokes</title><link
-    rel=preload as=font type=font/woff2 href=https://thecodinglove.com/wp-content/themes/thecodinglove/fonts/ibm-plex-sans-v7-latin-regular.woff2 crossorigin><link
-    rel=preload as=font type=font/woff2 href=https://thecodinglove.com/wp-content/themes/thecodinglove/fonts/ibm-plex-sans-v7-latin-500.woff2 crossorigin><link
-    rel=stylesheet href=https://thecodinglove.com/wp-content/cache/minify/66909.css media=all> <script async src=https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js></script>  <script async src="https://www.googletagmanager.com/gtag/js?id=UA-UA-31287421-2"></script> <script>window.dataLayer = window.dataLayer || [];
-    	  function gtag(){dataLayer.push(arguments);}
-    	  gtag('js', new Date());
-    	  gtag('config', 'UA-31287421-2');</script>  <script>(function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
-    	new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
-    	j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
-    	'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
-    	})(window,document,'script','dataLayer','GTM-WBTMH7V');</script> <meta
-    name='robots' content='noindex, follow'><meta
-    property="og:locale" content="en_US"><meta
-    property="og:type" content="article"><meta
-    property="og:title" content="You searched for test - The Coding Love - Programmer humor: gifs, memes, jokes"><meta
-    property="og:url" content="https://thecodinglove.com/search/test"><meta
-    property="og:site_name" content="The Coding Love - Programmer humor: gifs, memes, jokes"><meta
-    name="twitter:card" content="summary_large_image"><meta
-    name="twitter:title" content="You searched for test - The Coding Love - Programmer humor: gifs, memes, jokes"> <script type=application/ld+json class=yoast-schema-graph>{"@context":"https://schema.org","@graph":[{"@type":["CollectionPage","SearchResultsPage"],"@id":"https://thecodinglove.com/?s=test","url":"https://thecodinglove.com/?s=test","name":"You searched for test - The Coding Love - Programmer humor: gifs, memes, jokes","isPartOf":{"@id":"https://thecodinglove.com/#website"},"primaryImageOfPage":{"@id":"#primaryimage"},"image":{"@id":"#primaryimage"},"thumbnailUrl":"https://thecodinglove.com/wp-content/uploads/2022/11/Untitled-2-10.jpg","inLanguage":"en-US"},{"@type":"ImageObject","inLanguage":"en-US","@id":"#primaryimage","url":"https://thecodinglove.com/wp-content/uploads/2022/11/Untitled-2-10.jpg","contentUrl":"https://thecodinglove.com/wp-content/uploads/2022/11/Untitled-2-10.jpg","width":600,"height":300},{"@type":"WebSite","@id":"https://thecodinglove.com/#website","url":"https://thecodinglove.com/","name":"The Coding Love - Programmer humor: gifs, memes, jokes","description":"","potentialAction":[{"@type":"SearchAction","target":{"@type":"EntryPoint","urlTemplate":"https://thecodinglove.com/?s={search_term_string}"},"query-input":"required name=search_term_string"}],"inLanguage":"en-US"}]}</script> <link
-    rel=alternate type=application/rss+xml title="The Coding Love - Programmer humor: gifs, memes, jokes &raquo; Search Results for &#8220;test&#8221; Feed" href=https://thecodinglove.com/search/test/feed/rss2/ ><style id=global-styles-inline-css>/*<![CDATA[*/body{--wp--preset--color--black: #000000;--wp--preset--color--cyan-bluish-gray: #abb8c3;--wp--preset--color--white: #ffffff;--wp--preset--color--pale-pink: #f78da7;--wp--preset--color--vivid-red: #cf2e2e;--wp--preset--color--luminous-vivid-orange: #ff6900;--wp--preset--color--luminous-vivid-amber: #fcb900;--wp--preset--color--light-green-cyan: #7bdcb5;--wp--preset--color--vivid-green-cyan: #00d084;--wp--preset--color--pale-cyan-blue: #8ed1fc;--wp--preset--color--vivid-cyan-blue: #0693e3;--wp--preset--color--vivid-purple: #9b51e0;--wp--preset--gradient--vivid-cyan-blue-to-vivid-purple: linear-gradient(135deg,rgba(6,147,227,1) 0%,rgb(155,81,224) 100%);--wp--preset--gradient--light-green-cyan-to-vivid-green-cyan: linear-gradient(135deg,rgb(122,220,180) 0%,rgb(0,208,130) 100%);--wp--preset--gradient--luminous-vivid-amber-to-luminous-vivid-orange: linear-gradient(135deg,rgba(252,185,0,1) 0%,rgba(255,105,0,1) 100%);--wp--preset--gradient--luminous-vivid-orange-to-vivid-red: linear-gradient(135deg,rgba(255,105,0,1) 0%,rgb(207,46,46) 100%);--wp--preset--gradient--very-light-gray-to-cyan-bluish-gray: linear-gradient(135deg,rgb(238,238,238) 0%,rgb(169,184,195) 100%);--wp--preset--gradient--cool-to-warm-spectrum: linear-gradient(135deg,rgb(74,234,220) 0%,rgb(151,120,209) 20%,rgb(207,42,186) 40%,rgb(238,44,130) 60%,rgb(251,105,98) 80%,rgb(254,248,76) 100%);--wp--preset--gradient--blush-light-purple: linear-gradient(135deg,rgb(255,206,236) 0%,rgb(152,150,240) 100%);--wp--preset--gradient--blush-bordeaux: linear-gradient(135deg,rgb(254,205,165) 0%,rgb(254,45,45) 50%,rgb(107,0,62) 100%);--wp--preset--gradient--luminous-dusk: linear-gradient(135deg,rgb(255,203,112) 0%,rgb(199,81,192) 50%,rgb(65,88,208) 100%);--wp--preset--gradient--pale-ocean: linear-gradient(135deg,rgb(255,245,203) 0%,rgb(182,227,212) 50%,rgb(51,167,181) 100%);--wp--preset--gradient--electric-grass: linear-gradient(135deg,rgb(202,248,128) 0%,rgb(113,206,126) 100%);--wp--preset--gradient--midnight: linear-gradient(135deg,rgb(2,3,129) 0%,rgb(40,116,252) 100%);--wp--preset--duotone--dark-grayscale: url('#wp-duotone-dark-grayscale');--wp--preset--duotone--grayscale: url('#wp-duotone-grayscale');--wp--preset--duotone--purple-yellow: url('#wp-duotone-purple-yellow');--wp--preset--duotone--blue-red: url('#wp-duotone-blue-red');--wp--preset--duotone--midnight: url('#wp-duotone-midnight');--wp--preset--duotone--magenta-yellow: url('#wp-duotone-magenta-yellow');--wp--preset--duotone--purple-green: url('#wp-duotone-purple-green');--wp--preset--duotone--blue-orange: url('#wp-duotone-blue-orange');--wp--preset--font-size--small: 13px;--wp--preset--font-size--medium: 20px;--wp--preset--font-size--large: 36px;--wp--preset--font-size--x-large: 42px;--wp--preset--spacing--20: 0.44rem;--wp--preset--spacing--30: 0.67rem;--wp--preset--spacing--40: 1rem;--wp--preset--spacing--50: 1.5rem;--wp--preset--spacing--60: 2.25rem;--wp--preset--spacing--70: 3.38rem;--wp--preset--spacing--80: 5.06rem;}:where(.is-layout-flex){gap: 0.5em;}body .is-layout-flow > .alignleft{float: left;margin-inline-start: 0;margin-inline-end: 2em;}body .is-layout-flow > .alignright{float: right;margin-inline-start: 2em;margin-inline-end: 0;}body .is-layout-flow > .aligncenter{margin-left: auto !important;margin-right: auto !important;}body .is-layout-constrained > .alignleft{float: left;margin-inline-start: 0;margin-inline-end: 2em;}body .is-layout-constrained > .alignright{float: right;margin-inline-start: 2em;margin-inline-end: 0;}body .is-layout-constrained > .aligncenter{margin-left: auto !important;margin-right: auto !important;}body .is-layout-constrained > :where(:not(.alignleft):not(.alignright):not(.alignfull)){max-width: var(--wp--style--global--content-size);margin-left: auto !important;margin-right: auto !important;}body .is-layout-constrained > .alignwide{max-width: var(--wp--style--global--wide-size);}body .is-layout-flex{display: flex;}body .is-layout-flex{flex-wrap: wrap;align-items: center;}body .is-layout-flex > *{margin: 0;}:where(.wp-block-columns.is-layout-flex){gap: 2em;}.has-black-color{color: var(--wp--preset--color--black) !important;}.has-cyan-bluish-gray-color{color: var(--wp--preset--color--cyan-bluish-gray) !important;}.has-white-color{color: var(--wp--preset--color--white) !important;}.has-pale-pink-color{color: var(--wp--preset--color--pale-pink) !important;}.has-vivid-red-color{color: var(--wp--preset--color--vivid-red) !important;}.has-luminous-vivid-orange-color{color: var(--wp--preset--color--luminous-vivid-orange) !important;}.has-luminous-vivid-amber-color{color: var(--wp--preset--color--luminous-vivid-amber) !important;}.has-light-green-cyan-color{color: var(--wp--preset--color--light-green-cyan) !important;}.has-vivid-green-cyan-color{color: var(--wp--preset--color--vivid-green-cyan) !important;}.has-pale-cyan-blue-color{color: var(--wp--preset--color--pale-cyan-blue) !important;}.has-vivid-cyan-blue-color{color: var(--wp--preset--color--vivid-cyan-blue) !important;}.has-vivid-purple-color{color: var(--wp--preset--color--vivid-purple) !important;}.has-black-background-color{background-color: var(--wp--preset--color--black) !important;}.has-cyan-bluish-gray-background-color{background-color: var(--wp--preset--color--cyan-bluish-gray) !important;}.has-white-background-color{background-color: var(--wp--preset--color--white) !important;}.has-pale-pink-background-color{background-color: var(--wp--preset--color--pale-pink) !important;}.has-vivid-red-background-color{background-color: var(--wp--preset--color--vivid-red) !important;}.has-luminous-vivid-orange-background-color{background-color: var(--wp--preset--color--luminous-vivid-orange) !important;}.has-luminous-vivid-amber-background-color{background-color: var(--wp--preset--color--luminous-vivid-amber) !important;}.has-light-green-cyan-background-color{background-color: var(--wp--preset--color--light-green-cyan) !important;}.has-vivid-green-cyan-background-color{background-color: var(--wp--preset--color--vivid-green-cyan) !important;}.has-pale-cyan-blue-background-color{background-color: var(--wp--preset--color--pale-cyan-blue) !important;}.has-vivid-cyan-blue-background-color{background-color: var(--wp--preset--color--vivid-cyan-blue) !important;}.has-vivid-purple-background-color{background-color: var(--wp--preset--color--vivid-purple) !important;}.has-black-border-color{border-color: var(--wp--preset--color--black) !important;}.has-cyan-bluish-gray-border-color{border-color: var(--wp--preset--color--cyan-bluish-gray) !important;}.has-white-border-color{border-color: var(--wp--preset--color--white) !important;}.has-pale-pink-border-color{border-color: var(--wp--preset--color--pale-pink) !important;}.has-vivid-red-border-color{border-color: var(--wp--preset--color--vivid-red) !important;}.has-luminous-vivid-orange-border-color{border-color: var(--wp--preset--color--luminous-vivid-orange) !important;}.has-luminous-vivid-amber-border-color{border-color: var(--wp--preset--color--luminous-vivid-amber) !important;}.has-light-green-cyan-border-color{border-color: var(--wp--preset--color--light-green-cyan) !important;}.has-vivid-green-cyan-border-color{border-color: var(--wp--preset--color--vivid-green-cyan) !important;}.has-pale-cyan-blue-border-color{border-color: var(--wp--preset--color--pale-cyan-blue) !important;}.has-vivid-cyan-blue-border-color{border-color: var(--wp--preset--color--vivid-cyan-blue) !important;}.has-vivid-purple-border-color{border-color: var(--wp--preset--color--vivid-purple) !important;}.has-vivid-cyan-blue-to-vivid-purple-gradient-background{background: var(--wp--preset--gradient--vivid-cyan-blue-to-vivid-purple) !important;}.has-light-green-cyan-to-vivid-green-cyan-gradient-background{background: var(--wp--preset--gradient--light-green-cyan-to-vivid-green-cyan) !important;}.has-luminous-vivid-amber-to-luminous-vivid-orange-gradient-background{background: var(--wp--preset--gradient--luminous-vivid-amber-to-luminous-vivid-orange) !important;}.has-luminous-vivid-orange-to-vivid-red-gradient-background{background: var(--wp--preset--gradient--luminous-vivid-orange-to-vivid-red) !important;}.has-very-light-gray-to-cyan-bluish-gray-gradient-background{background: var(--wp--preset--gradient--very-light-gray-to-cyan-bluish-gray) !important;}.has-cool-to-warm-spectrum-gradient-background{background: var(--wp--preset--gradient--cool-to-warm-spectrum) !important;}.has-blush-light-purple-gradient-background{background: var(--wp--preset--gradient--blush-light-purple) !important;}.has-blush-bordeaux-gradient-background{background: var(--wp--preset--gradient--blush-bordeaux) !important;}.has-luminous-dusk-gradient-background{background: var(--wp--preset--gradient--luminous-dusk) !important;}.has-pale-ocean-gradient-background{background: var(--wp--preset--gradient--pale-ocean) !important;}.has-electric-grass-gradient-background{background: var(--wp--preset--gradient--electric-grass) !important;}.has-midnight-gradient-background{background: var(--wp--preset--gradient--midnight) !important;}.has-small-font-size{font-size: var(--wp--preset--font-size--small) !important;}.has-medium-font-size{font-size: var(--wp--preset--font-size--medium) !important;}.has-large-font-size{font-size: var(--wp--preset--font-size--large) !important;}.has-x-large-font-size{font-size: var(--wp--preset--font-size--x-large) !important;}
-    .wp-block-navigation a:where(:not(.wp-element-button)){color: inherit;}
-    :where(.wp-block-columns.is-layout-flex){gap: 2em;}
-    .wp-block-pullquote{font-size: 1.5em;line-height: 1.6;}/*]]>*/</style><link
-    rel=https://api.w.org/ href=https://thecodinglove.com/wp-json/ ><link
-    rel=EditURI type=application/rsd+xml title=RSD href=https://thecodinglove.com/xmlrpc.php?rsd><link
-    rel=wlwmanifest type=application/wlwmanifest+xml href=https://thecodinglove.com/wp-includes/wlwmanifest.xml><meta
-    name="generator" content="WordPress 6.1.1"><style>img#wpstats{display:none}</style> <script>!function(f,b,e,v,n,t,s){if(f.fbq)return;n=f.fbq=function(){n.callMethod?
-    n.callMethod.apply(n,arguments):n.queue.push(arguments)};if(!f._fbq)f._fbq=n;
-    n.push=n;n.loaded=!0;n.version='2.0';n.queue=[];t=b.createElement(e);t.async=!0;
-    t.src=v;s=b.getElementsByTagName(e)[0];s.parentNode.insertBefore(t,s)}(window,
-    document,'script','https://connect.facebook.net/en_US/fbevents.js');</script>  <script>fbq('init', '247014279189106', {}, {
-        "agent": "wordpress-6.1.1-3.0.8"
-    });</script><script>fbq('track', 'PageView', []);</script> 
-    <noscript>
-    <img
-    class=lazy height=1 width=1 style=display:none alt=fbpx
-    src="data:image/svg+xml,%3Csvg%20xmlns='http://www.w3.org/2000/svg'%20viewBox='0%200%201%201'%3E%3C/svg%3E" data-src="https://www.facebook.com/tr?id=247014279189106&ev=PageView&noscript=1">
-    </noscript><link
-    rel=icon href=https://thecodinglove.com/wp-content/uploads/2020/03/cropped-59760110_2118870124856761_2769282087964901376_n-32x32.png sizes=32x32><link
-    rel=icon href=https://thecodinglove.com/wp-content/uploads/2020/03/cropped-59760110_2118870124856761_2769282087964901376_n-192x192.png sizes=192x192><link
-    rel=apple-touch-icon href=https://thecodinglove.com/wp-content/uploads/2020/03/cropped-59760110_2118870124856761_2769282087964901376_n-180x180.png><meta
-    name="msapplication-TileImage" content="https://thecodinglove.com/wp-content/uploads/2020/03/cropped-59760110_2118870124856761_2769282087964901376_n-270x270.png"></head><body>
-    <noscript><iframe
-    src="https://www.googletagmanager.com/ns.html?id=GTM-WBTMH7V"
-    height=0 width=0 style=display:none;visibility:hidden></iframe></noscript><nav
-    class="navbar navbar-expand-xl navbar-light">
-    <a
-    class=navbar-brand href=https://thecodinglove.com>
-    <svg
-    class=brand-name
-    xmlns:dc=http://purl.org/dc/elements/1.1/
-    xmlns:cc=http://creativecommons.org/ns#
-    xmlns:rdf=http://www.w3.org/1999/02/22-rdf-syntax-ns#
-    xmlns:svg=http://www.w3.org/2000/svg
-    xmlns=http://www.w3.org/2000/svg
-    viewBox="0 0 153.10121 19.243401">
-    <path
-    style=fill:#000000;stroke-width:0.21892475
-    d="M 3.1417804,0.02048639 H 127.3668 c 1.74054,0 3.14178,1.40123401 3.14178,3.14178041 V 16.122107 c 0,1.740546 -1.40124,3.14178 -3.14178,3.14178 H 3.1417804 C 1.401234,19.263887 0,17.862653 0,16.122107 V 3.1622668 C 0,1.4217204 1.401234,0.02048639 3.1417804,0.02048639 Z"
-    id=rect3719 />
-    <g
-    aria-label="the_coding_love ();"
-    style="font-style:normal;font-variant:normal;font-weight:normal;font-stretch:normal;font-size:14.11111069px;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code';letter-spacing:0px;word-spacing:0px;fill:#000000;fill-opacity:1;stroke:none;stroke-width:0.26458332"
-    id=text3715>
-    <path
-    d="M 9.1359629,14.185326 8.5715184,13.141103 c -0.3809999,0.211667 -0.8466666,0.352778 -1.3828888,0.352778 -0.9172222,0 -1.2982222,-0.381 -1.2982222,-1.185333 V 8.2868813 H 8.2328518 L 8.4162962,7.1297703 H 5.8904074 V 5.2812148 L 4.309963,5.4646592 V 7.1297703 H 2.6589631 v 1.157111 H 4.309963 v 4.0357777 c 0,1.467556 0.9172222,2.441222 2.6528888,2.441222 0.8325556,0 1.6227778,-0.225778 2.1731111,-0.578555 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#ffffff;stroke-width:0.26458332"
-    id=path3822 />
-    <path
-    d="M 12.954561,3.9829926 11.374117,4.1523259 V 14.580437 h 1.580444 V 9.2464369 c 0.395112,-0.5644445 1.030112,-1.1006667 1.735667,-1.1006667 0.747889,0 1.072444,0.3527778 1.072444,1.2417778 v 5.192889 h 1.580445 V 9.1476591 c 0,-1.397 -0.677333,-2.2154444 -2.074333,-2.2154444 -0.874889,0 -1.707445,0.381 -2.314223,1.1147778 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#ffffff;stroke-width:0.26458332"
-    id=path3824 />
-    <path
-    d="m 21.147604,11.363103 h 4.938889 c 0.02822,-0.183444 0.04233,-0.451555 0.04233,-0.705555 0,-2.3142222 -1.241777,-3.7253333 -3.217333,-3.7253333 -2.116666,0 -3.400778,1.6933333 -3.400778,3.9228883 0,2.328334 1.27,3.908778 3.598334,3.908778 1.044222,0 2.017889,-0.352778 2.709333,-0.917222 l -0.677333,-0.959556 c -0.635,0.423334 -1.241778,0.649112 -1.961445,0.649112 -1.072444,0 -1.933222,-0.649112 -2.032,-2.173112 z m 1.778,-3.2596661 c 1.058334,0 1.636889,0.7761111 1.665111,2.1872221 h -3.457222 c 0.09878,-1.4957777 0.804334,-2.1872221 1.792111,-2.1872221 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#ffffff;stroke-width:0.26458332"
-    id=path3826 />
-    <path
-    d="m 27.746091,17.16277 h 7.055555 v -1.298222 h -7.055555 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#ffffff;stroke-width:0.26458332"
-    id=path3828 />
-    <path
-    d="m 40.355912,13.465659 c -1.157111,0 -2.003777,-0.705556 -2.003777,-2.568222 0,-1.8626668 0.818444,-2.6952223 2.003777,-2.6952223 0.635,0 1.241778,0.2257777 1.820334,0.6208888 L 42.910023,7.8353258 C 42.190357,7.2426591 41.34369,6.9322147 40.32769,6.9322147 c -2.300111,0 -3.654778,1.6933333 -3.654778,3.9652223 0,2.328333 1.326445,3.866444 3.626556,3.866444 0.987778,0 1.961444,-0.366889 2.610555,-0.917222 l -0.719666,-1.016 c -0.550334,0.366889 -1.143,0.635 -1.834445,0.635 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#ffffff;stroke-width:0.26458332"
-    id=path3830 />
-    <path
-    d="m 48.210289,6.9322147 c -2.173111,0 -3.386666,1.5945555 -3.386666,3.9228883 0,2.413 1.199444,3.908778 3.372555,3.908778 2.159,0 3.372556,-1.552222 3.372556,-3.922889 0,-2.3706662 -1.185334,-3.9087773 -3.358445,-3.9087773 z m 0,1.2276666 c 1.128889,0 1.693334,0.8466667 1.693334,2.6811107 0,1.862667 -0.564445,2.709334 -1.707445,2.709334 -1.143,0 -1.707444,-0.846667 -1.707444,-2.695223 0,-1.848555 0.578555,-2.6952217 1.721555,-2.6952217 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#ffffff;stroke-width:0.26458332"
-    id=path3832 />
-    <path
-    d="M 58.054329,3.9547704 V 7.7506591 C 57.574552,7.2144369 56.953663,6.9322147 56.177552,6.9322147 c -1.834445,0 -2.949222,1.651 -2.949222,3.9228883 0,2.342445 0.888999,3.908778 2.808111,3.908778 0.874888,0 1.594555,-0.395111 2.088444,-1.100666 l 0.112889,0.917222 h 1.397 V 4.1382148 Z m -1.467555,4.1909998 c 0.606778,0 1.100666,0.3104445 1.467555,0.8325556 v 3.5136662 c -0.366889,0.592667 -0.889,1.072445 -1.594555,1.072445 -0.987778,0 -1.552222,-0.804334 -1.552222,-2.709334 0,-1.8626661 0.620889,-2.7093328 1.679222,-2.7093328 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#ffffff;stroke-width:0.26458332"
-    id=path3834 />
-    <path
-    d="m 65.118484,3.4749926 c -0.620889,0 -1.030111,0.4374444 -1.030111,1.0018889 0,0.5644444 0.409222,0.9877777 1.030111,0.9877777 0.606778,0 1.030111,-0.4233333 1.030111,-0.9877777 0,-0.5644445 -0.423333,-1.0018889 -1.030111,-1.0018889 z m 1.086556,3.6547777 h -3.725333 v 1.157111 h 2.144888 v 5.1364447 h -2.215444 v 1.157111 h 5.827889 v -1.157111 h -2.032 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#ffffff;stroke-width:0.26458332"
-    id=path3836 />
-    <path
-    d="m 70.602195,14.580437 h 1.580444 V 9.2323258 c 0.395111,-0.5644445 1.016,-1.0865556 1.749778,-1.0865556 0.903111,0 1.058333,0.4938889 1.058333,1.7921111 v 4.6425557 h 1.580445 V 9.1476591 c 0,-1.397 -0.691445,-2.2154444 -2.074334,-2.2154444 -0.903111,0 -1.792111,0.4092222 -2.412999,1.1712222 L 71.970973,7.1297703 h -1.368778 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#ffffff;stroke-width:0.26458332"
-    id=path3838 />
-    <path
-    d="m 85.34279,6.1843258 c -0.889,0.4374445 -1.651,0.762 -3.443111,0.7478889 -1.834444,0 -3.104444,1.0724444 -3.104444,2.6387777 0,0.9736666 0.381,1.6227776 1.27,2.0884446 -0.536222,0.352778 -0.818445,0.846666 -0.818445,1.340555 0,0.776111 0.620889,1.439334 2.032,1.439334 h 1.284112 c 0.874888,0 1.439333,0.338666 1.439333,0.945444 0,0.691444 -0.564445,1.086556 -2.116667,1.086556 -1.566333,0 -1.919111,-0.366889 -1.919111,-1.157112 h -1.411111 c 0,1.509889 0.747889,2.314223 3.316111,2.314223 2.413,0 3.725333,-0.860778 3.725333,-2.342445 0,-1.199444 -1.128889,-2.074333 -2.723444,-2.074333 h -1.298222 c -0.747889,0 -0.917222,-0.268111 -0.917222,-0.564444 0,-0.239889 0.127,-0.479778 0.324555,-0.620889 0.296333,0.09878 0.592667,0.127 1.016,0.127 1.834444,0 2.949222,-1.058334 2.949222,-2.511778 0,-0.9031111 -0.465666,-1.4957778 -1.354666,-1.905 0.874888,0 1.580444,-0.042333 2.159,-0.2398889 z m -3.429,1.8485555 c 0.973667,0 1.453445,0.508 1.453445,1.5098889 0,1.0159998 -0.493889,1.5804448 -1.439333,1.5804448 -1.016,0 -1.524,-0.620889 -1.524,-1.5522226 0,-0.9172222 0.493888,-1.5381111 1.509888,-1.5381111 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#ffffff;stroke-width:0.26458332"
-    id=path3840 />
-    <path
-    d="m 86.974171,17.16277 h 7.055556 v -1.298222 h -7.055556 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#ffffff;stroke-width:0.26458332"
-    id=path3842 />
-    <path
-    d="m 99.259434,4.1382148 h -3.824111 v 1.1571111 h 2.243666 v 7.1120001 c 0,1.467555 0.874889,2.356555 2.370671,2.356555 0.83255,0 1.48166,-0.183444 1.91911,-0.423333 l -0.39511,-1.100667 c -0.381,0.155222 -0.762,0.254 -1.18534,0.254 -0.66322,0 -1.128886,-0.254 -1.128886,-0.945444 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#ffffff;stroke-width:0.26458332"
-    id=path3844 />
-    <path
-    d="m 107.43836,6.9322147 c -2.17311,0 -3.38666,1.5945555 -3.38666,3.9228883 0,2.413 1.19944,3.908778 3.37255,3.908778 2.159,0 3.37256,-1.552222 3.37256,-3.922889 0,-2.3706662 -1.18534,-3.9087773 -3.35845,-3.9087773 z m 0,1.2276666 c 1.12889,0 1.69334,0.8466667 1.69334,2.6811107 0,1.862667 -0.56445,2.709334 -1.70745,2.709334 -1.143,0 -1.70744,-0.846667 -1.70744,-2.695223 0,-1.848555 0.57855,-2.6952217 1.72155,-2.6952217 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#ffffff;stroke-width:0.26458332"
-    id=path3846 />
-    <path
-    d="m 119.44141,7.1297703 h -1.62278 l -1.905,6.1383327 -1.87678,-6.1383327 h -1.70744 l 2.63878,7.4506667 h 1.83444 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#ffffff;stroke-width:0.26458332"
-    id=path3848 />
-    <path
-    d="m 122.68145,11.363103 h 4.93889 c 0.0282,-0.183444 0.0423,-0.451555 0.0423,-0.705555 0,-2.3142222 -1.24178,-3.7253333 -3.21733,-3.7253333 -2.11667,0 -3.40078,1.6933333 -3.40078,3.9228883 0,2.328334 1.27,3.908778 3.59833,3.908778 1.04423,0 2.01789,-0.352778 2.70934,-0.917222 l -0.67734,-0.959556 c -0.635,0.423334 -1.24177,0.649112 -1.96144,0.649112 -1.07244,0 -1.93322,-0.649112 -2.032,-2.173112 z m 1.778,-3.2596661 c 1.05833,0 1.63689,0.7761111 1.66511,2.1872221 h -3.45722 c 0.0988,-1.4957777 0.80433,-2.1872221 1.79211,-2.1872221 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#ffffff;stroke-width:0.26458332"
-    id=path3850 />
-    <path
-    d="m 133.94933,9.641548 c 0,-2.7234444 0.90311,-4.2615555 2.87867,-6.1665554 L 135.96722,2.628326 c -2.27189,1.8908888 -3.64066,3.8382221 -3.64066,7.013222 0,3.175 1.36877,5.122333 3.64066,7.013222 l 0.86078,-0.846667 c -2.01789,-1.947333 -2.87867,-3.457222 -2.87867,-6.166555 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#000000;stroke-width:0.26458332"
-    id=path3852 />
-    <path
-    d="m 144.35783,9.641548 c 0,2.709333 -0.86077,4.219222 -2.87866,6.166555 l 0.86077,0.846667 c 2.27189,-1.890889 3.64067,-3.838222 3.64067,-7.013222 0,-3.1749999 -1.36878,-5.1223332 -3.64067,-7.013222 l -0.86077,0.8466666 c 1.97555,1.9049999 2.87866,3.443111 2.87866,6.1665554 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#000000;stroke-width:0.26458332"
-    id=path3854 />
-    <path
-    d="m 151.83121,7.2144369 c -0.70556,0 -1.25589,0.5785556 -1.25589,1.27 0,0.6914444 0.55033,1.27 1.25589,1.27 0.73377,0 1.27,-0.5785556 1.27,-1.27 0,-0.6914444 -0.53623,-1.27 -1.27,-1.27 z m 0,4.8542221 c -0.71967,0 -1.25589,0.564444 -1.25589,1.27 0,0.423333 0.19755,0.818444 0.52211,1.044222 l -0.67733,2.836333 h 1.21355 l 1.08656,-2.483555 c 0.26811,-0.606778 0.381,-0.987778 0.381,-1.397 0,-0.705556 -0.53623,-1.27 -1.27,-1.27 z"
-    style="font-style:normal;font-variant:normal;font-weight:500;font-stretch:normal;line-height:1.25;font-family:'Fira Code';-inkscape-font-specification:'Fira Code Medium';fill:#000000;stroke-width:0.26458332"
-    id=path3856 />
-    </g>
-    </svg>
-    </a><button
-    class="navbar-toggler custom-toggler" type=button data-toggle=collapse data-target=#mainNavbar aria-controls=mainNavbar aria-expanded=false aria-label="Toggle navigation">
-    <span
-    class=navbar-toggler-icon></span>
-    </button><div
-    class="collapse navbar-collapse" id=mainNavbar><ul
-    class="navbar-nav mr-auto">
-    <li
-    class=nav-item>
-    <a
-    class=nav-link href=https://thecodinglove.com/random>
-    <svg
-    class=nav-link-icon aria-hidden=true focusable=false data-prefix=fas data-icon=dice class="svg-inline--fa fa-dice fa-w-20" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 640 512"><path
-    fill=currentColor d="M592 192H473.26c12.69 29.59 7.12 65.2-17 89.32L320 417.58V464c0 26.51 21.49 48 48 48h224c26.51 0 48-21.49 48-48V240c0-26.51-21.49-48-48-48zM480 376c-13.25 0-24-10.75-24-24 0-13.26 10.75-24 24-24s24 10.74 24 24c0 13.25-10.75 24-24 24zm-46.37-186.7L258.7 14.37c-19.16-19.16-50.23-19.16-69.39 0L14.37 189.3c-19.16 19.16-19.16 50.23 0 69.39L189.3 433.63c19.16 19.16 50.23 19.16 69.39 0L433.63 258.7c19.16-19.17 19.16-50.24 0-69.4zM96 248c-13.25 0-24-10.75-24-24 0-13.26 10.75-24 24-24s24 10.74 24 24c0 13.25-10.75 24-24 24zm128 128c-13.25 0-24-10.75-24-24 0-13.26 10.75-24 24-24s24 10.74 24 24c0 13.25-10.75 24-24 24zm0-128c-13.25 0-24-10.75-24-24 0-13.26 10.75-24 24-24s24 10.74 24 24c0 13.25-10.75 24-24 24zm0-128c-13.25 0-24-10.75-24-24 0-13.26 10.75-24 24-24s24 10.74 24 24c0 13.25-10.75 24-24 24zm128 128c-13.25 0-24-10.75-24-24 0-13.26 10.75-24 24-24s24 10.74 24 24c0 13.25-10.75 24-24 24z"></path></svg>
-    random</a>
-    </li></ul><ul
-    class="navbar-nav ml-auto"><form
-    class="navbar-form search-form" role=search onsubmit="location.href='https://thecodinglove.com/?s=' + document.getElementById('search-input').value; return false;"><div
-    class=input-group>
-    <input
-    id=search-input name=s type=search class="form-control search-input" placeholder=Search><div
-    class=input-group-btn>
-    <button
-    class="btn btn-default search-btn" type=submit>
-    <svg
-    class=search-icon aria-hidden=true focusable=false data-prefix=fas data-icon=search class="svg-inline--fa fa-search fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M505 442.7L405.3 343c-4.5-4.5-10.6-7-17-7H372c27.6-35.3 44-79.7 44-128C416 93.1 322.9 0 208 0S0 93.1 0 208s93.1 208 208 208c48.3 0 92.7-16.4 128-44v16.3c0 6.4 2.5 12.5 7 17l99.7 99.7c9.4 9.4 24.6 9.4 33.9 0l28.3-28.3c9.4-9.4 9.4-24.6.1-34zM208 336c-70.7 0-128-57.2-128-128 0-70.7 57.2-128 128-128 70.7 0 128 57.2 128 128 0 70.7-57.2 128-128 128z"></path></svg>
-    </button></div></div></form>
-    <a
-    href=https://thecodinglove.com/submit>
-    <button
-    id=submitPostBtn class="btn my-2 my-sm-0
-    submit-post-btn					" 					>
-    <svg
-    class=submit-icon aria-hidden=true focusable=false data-prefix=fas data-icon=plus class="svg-inline--fa fa-plus fa-w-14" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 448 512"><path
-    fill=currentColor d="M416 208H272V64c0-17.67-14.33-32-32-32h-32c-17.67 0-32 14.33-32 32v144H32c-17.67 0-32 14.33-32 32v32c0 17.67 14.33 32 32 32h144v144c0 17.67 14.33 32 32 32h32c17.67 0 32-14.33 32-32V304h144c17.67 0 32-14.33 32-32v-32c0-17.67-14.33-32-32-32z"></path></svg>
-    Submit post
-    </button>
-    </a></ul></div></nav><main
-    role=main class=container><div
-    class=row><div
-    class="col-md-8 blog-main"><div
-    class="blog-post results-return"><span
-    class=results-number>314</span> results found for: <span
-    class=search-text>test</span></div><article
-    class="blog-post index-blog-post "><h1 class="index-blog-post-title">
-    <a
-    href=https://thecodinglove.com/when-the-sales-guy-is-about-to-click-on-a-button-i-never-tested>When the sales guy is about to click on a button I never tested</a></h1><div
-    class=post-meta-info>
-    <svg
-    class=octicon-svg xmlns=http://www.w3.org/2000/svg width=14 height=16 viewBox="0 0 14 16"><path
-    fill-rule=evenodd d="M10.86 7c-.45-1.72-2-3-3.86-3-1.86 0-3.41 1.28-3.86 3H0v2h3.14c.45 1.72 2 3 3.86 3 1.86 0 3.41-1.28 3.86-3H14V7h-3.14zM7 10.2c-1.22 0-2.2-.98-2.2-2.2 0-1.22.98-2.2 2.2-2.2 1.22 0 2.2.98 2.2 2.2 0 1.22-.98 2.2-2.2 2.2z"/></svg>
-    <span
-    class=post-author>The Coding Love</span> commited on 16 Nov 2022</div><div
-    class=blog-post-content><p><video
-    autoplay loop muted playsinline style=width:400px;><source
-    src=https://thecodinglove.com/content/047/6chNSPh.webm type=video/webm><source
-    src=https://thecodinglove.com/content/047/6chNSPh.mp4 type=video/mp4><object
-    data=https://thecodinglove.com/content/047/6chNSPh.gif type=image/gif></object></video></p></div><div
-    class=social-share><a
-    href="https://www.facebook.com/sharer/sharer.php?u=https://thecodinglove.com/when-the-sales-guy-is-about-to-click-on-a-button-i-never-tested" class="js-social-share social-btn" target=_blank>
-    <button
-    id=facebookSharingBtn type=button class="btn btn-default share-fb-btn">
-    <svg
-    class=sharing-btn-icon aria-hidden=true focusable=false data-prefix=fab data-icon=facebook class="svg-inline--fa fa-facebook fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M504 256C504 119 393 8 256 8S8 119 8 256c0 123.78 90.69 226.38 209.25 245V327.69h-63V256h63v-54.64c0-62.15 37-96.48 93.67-96.48 27.14 0 55.52 4.84 55.52 4.84v61h-31.28c-30.8 0-40.41 19.12-40.41 38.73V256h68.78l-11 71.69h-57.78V501C413.31 482.38 504 379.78 504 256z"></path></svg>
-    Facebook
-    </button>
-    </a>
-    <a
-    href="https://twitter.com/intent/tweet?text=When the sales guy is about to click on a button I never tested&url=https://thecodinglove.com/when-the-sales-guy-is-about-to-click-on-a-button-i-never-tested&related=thecodinglove&hashtags=thecodinglove" class=social-btn target=_blank>
-    <button
-    id=twitterSharingBtn type=button class="btn btn-default share-twitter-btn">
-    <svg
-    class=sharing-btn-icon aria-hidden=true focusable=false data-prefix=fab data-icon=twitter class="svg-inline--fa fa-twitter fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M459.37 151.716c.325 4.548.325 9.097.325 13.645 0 138.72-105.583 298.558-298.558 298.558-59.452 0-114.68-17.219-161.137-47.106 8.447.974 16.568 1.299 25.34 1.299 49.055 0 94.213-16.568 130.274-44.832-46.132-.975-84.792-31.188-98.112-72.772 6.498.974 12.995 1.624 19.818 1.624 9.421 0 18.843-1.3 27.614-3.573-48.081-9.747-84.143-51.98-84.143-102.985v-1.299c13.969 7.797 30.214 12.67 47.431 13.319-28.264-18.843-46.781-51.005-46.781-87.391 0-19.492 5.197-37.36 14.294-52.954 51.655 63.675 129.3 105.258 216.365 109.807-1.624-7.797-2.599-15.918-2.599-24.04 0-57.828 46.782-104.934 104.934-104.934 30.213 0 57.502 12.67 76.67 33.137 23.715-4.548 46.456-13.32 66.599-25.34-7.798 24.366-24.366 44.833-46.132 57.827 21.117-2.273 41.584-8.122 60.426-16.243-14.292 20.791-32.161 39.308-52.628 54.253z"></path></svg>
-    Twitter
-    </button>
-    </a><a
-    href="mailto:?body=https://thecodinglove.com/when-the-sales-guy-is-about-to-click-on-a-button-i-never-tested&subject=" class=social-btn>
-    <button
-    id=emailSharingBtn type=button class="btn btn-default share-email-btn">
-    <svg
-    class=sharing-btn-icon aria-hidden=true focusable=false data-prefix=fas data-icon=envelope class="svg-inline--fa fa-envelope fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M502.3 190.8c3.9-3.1 9.7-.2 9.7 4.7V400c0 26.5-21.5 48-48 48H48c-26.5 0-48-21.5-48-48V195.6c0-5 5.7-7.8 9.7-4.7 22.4 17.4 52.1 39.5 154.1 113.6 21.1 15.4 56.7 47.8 92.2 47.6 35.7.3 72-32.8 92.3-47.6 102-74.1 131.6-96.3 154-113.7zM256 320c23.2.4 56.6-29.2 73.4-41.4 132.7-96.3 142.8-104.7 173.4-128.7 5.8-4.5 9.2-11.5 9.2-18.9v-19c0-26.5-21.5-48-48-48H48C21.5 64 0 85.5 0 112v19c0 7.4 3.4 14.3 9.2 18.9 30.6 23.9 40.7 32.4 173.4 128.7 16.8 12.2 50.2 41.8 73.4 41.4z"></path></svg>
-    Email
-    </button>
-    </a>
-    <button
-    id=copyLinkBtn type=button title="Copy link" class="btn btn-default share-link-btn" data-clipboard-text=https://thecodinglove.com/when-the-sales-guy-is-about-to-click-on-a-button-i-never-tested data-toggle=tooltip data-placement=right>
-    <svg
-    class=sharing-btn-icon aria-hidden=true focusable=false data-prefix=fas data-icon=link class="svg-inline--fa fa-link fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M326.612 185.391c59.747 59.809 58.927 155.698.36 214.59-.11.12-.24.25-.36.37l-67.2 67.2c-59.27 59.27-155.699 59.262-214.96 0-59.27-59.26-59.27-155.7 0-214.96l37.106-37.106c9.84-9.84 26.786-3.3 27.294 10.606.648 17.722 3.826 35.527 9.69 52.721 1.986 5.822.567 12.262-3.783 16.612l-13.087 13.087c-28.026 28.026-28.905 73.66-1.155 101.96 28.024 28.579 74.086 28.749 102.325.51l67.2-67.19c28.191-28.191 28.073-73.757 0-101.83-3.701-3.694-7.429-6.564-10.341-8.569a16.037 16.037 0 0 1-6.947-12.606c-.396-10.567 3.348-21.456 11.698-29.806l21.054-21.055c5.521-5.521 14.182-6.199 20.584-1.731a152.482 152.482 0 0 1 20.522 17.197zM467.547 44.449c-59.261-59.262-155.69-59.27-214.96 0l-67.2 67.2c-.12.12-.25.25-.36.37-58.566 58.892-59.387 154.781.36 214.59a152.454 152.454 0 0 0 20.521 17.196c6.402 4.468 15.064 3.789 20.584-1.731l21.054-21.055c8.35-8.35 12.094-19.239 11.698-29.806a16.037 16.037 0 0 0-6.947-12.606c-2.912-2.005-6.64-4.875-10.341-8.569-28.073-28.073-28.191-73.639 0-101.83l67.2-67.19c28.239-28.239 74.3-28.069 102.325.51 27.75 28.3 26.872 73.934-1.155 101.96l-13.087 13.087c-4.35 4.35-5.769 10.79-3.783 16.612 5.864 17.194 9.042 34.999 9.69 52.721.509 13.906 17.454 20.446 27.294 10.606l37.106-37.106c59.271-59.259 59.271-155.699.001-214.959z"></path></svg>
-    </button></div></article><article
-    class="blog-post index-blog-post "><h1 class="index-blog-post-title">
-    <a
-    href=https://thecodinglove.com/when-my-coworker-asks-me-if-i-can-test-his-code>When my coworker asks me if I can test his code</a></h1><div
-    class=post-meta-info>
-    <svg
-    class=octicon-svg xmlns=http://www.w3.org/2000/svg width=14 height=16 viewBox="0 0 14 16"><path
-    fill-rule=evenodd d="M10.86 7c-.45-1.72-2-3-3.86-3-1.86 0-3.41 1.28-3.86 3H0v2h3.14c.45 1.72 2 3 3.86 3 1.86 0 3.41-1.28 3.86-3H14V7h-3.14zM7 10.2c-1.22 0-2.2-.98-2.2-2.2 0-1.22.98-2.2 2.2-2.2 1.22 0 2.2.98 2.2 2.2 0 1.22-.98 2.2-2.2 2.2z"/></svg>
-    <span
-    class=post-author>The Coding Love</span> commited on 14 Nov 2022</div><div
-    class=blog-post-content><p><video
-    autoplay loop muted playsinline><source
-    src=https://thecodinglove.com/content/047/hk9fPUH.webm type=video/webm><source
-    src=https://thecodinglove.com/content/047/hk9fPUH.mp4 type=video/mp4><object
-    data=https://thecodinglove.com/content/047/hk9fPUH.gif type=image/gif></object></video></p></div><div
-    class=social-share><a
-    href="https://www.facebook.com/sharer/sharer.php?u=https://thecodinglove.com/when-my-coworker-asks-me-if-i-can-test-his-code" class="js-social-share social-btn" target=_blank>
-    <button
-    id=facebookSharingBtn type=button class="btn btn-default share-fb-btn">
-    <svg
-    class=sharing-btn-icon aria-hidden=true focusable=false data-prefix=fab data-icon=facebook class="svg-inline--fa fa-facebook fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M504 256C504 119 393 8 256 8S8 119 8 256c0 123.78 90.69 226.38 209.25 245V327.69h-63V256h63v-54.64c0-62.15 37-96.48 93.67-96.48 27.14 0 55.52 4.84 55.52 4.84v61h-31.28c-30.8 0-40.41 19.12-40.41 38.73V256h68.78l-11 71.69h-57.78V501C413.31 482.38 504 379.78 504 256z"></path></svg>
-    Facebook
-    </button>
-    </a>
-    <a
-    href="https://twitter.com/intent/tweet?text=When my coworker asks me if I can test his code&url=https://thecodinglove.com/when-my-coworker-asks-me-if-i-can-test-his-code&related=thecodinglove&hashtags=thecodinglove" class=social-btn target=_blank>
-    <button
-    id=twitterSharingBtn type=button class="btn btn-default share-twitter-btn">
-    <svg
-    class=sharing-btn-icon aria-hidden=true focusable=false data-prefix=fab data-icon=twitter class="svg-inline--fa fa-twitter fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M459.37 151.716c.325 4.548.325 9.097.325 13.645 0 138.72-105.583 298.558-298.558 298.558-59.452 0-114.68-17.219-161.137-47.106 8.447.974 16.568 1.299 25.34 1.299 49.055 0 94.213-16.568 130.274-44.832-46.132-.975-84.792-31.188-98.112-72.772 6.498.974 12.995 1.624 19.818 1.624 9.421 0 18.843-1.3 27.614-3.573-48.081-9.747-84.143-51.98-84.143-102.985v-1.299c13.969 7.797 30.214 12.67 47.431 13.319-28.264-18.843-46.781-51.005-46.781-87.391 0-19.492 5.197-37.36 14.294-52.954 51.655 63.675 129.3 105.258 216.365 109.807-1.624-7.797-2.599-15.918-2.599-24.04 0-57.828 46.782-104.934 104.934-104.934 30.213 0 57.502 12.67 76.67 33.137 23.715-4.548 46.456-13.32 66.599-25.34-7.798 24.366-24.366 44.833-46.132 57.827 21.117-2.273 41.584-8.122 60.426-16.243-14.292 20.791-32.161 39.308-52.628 54.253z"></path></svg>
-    Twitter
-    </button>
-    </a><a
-    href="mailto:?body=https://thecodinglove.com/when-my-coworker-asks-me-if-i-can-test-his-code&subject=" class=social-btn>
-    <button
-    id=emailSharingBtn type=button class="btn btn-default share-email-btn">
-    <svg
-    class=sharing-btn-icon aria-hidden=true focusable=false data-prefix=fas data-icon=envelope class="svg-inline--fa fa-envelope fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M502.3 190.8c3.9-3.1 9.7-.2 9.7 4.7V400c0 26.5-21.5 48-48 48H48c-26.5 0-48-21.5-48-48V195.6c0-5 5.7-7.8 9.7-4.7 22.4 17.4 52.1 39.5 154.1 113.6 21.1 15.4 56.7 47.8 92.2 47.6 35.7.3 72-32.8 92.3-47.6 102-74.1 131.6-96.3 154-113.7zM256 320c23.2.4 56.6-29.2 73.4-41.4 132.7-96.3 142.8-104.7 173.4-128.7 5.8-4.5 9.2-11.5 9.2-18.9v-19c0-26.5-21.5-48-48-48H48C21.5 64 0 85.5 0 112v19c0 7.4 3.4 14.3 9.2 18.9 30.6 23.9 40.7 32.4 173.4 128.7 16.8 12.2 50.2 41.8 73.4 41.4z"></path></svg>
-    Email
-    </button>
-    </a>
-    <button
-    id=copyLinkBtn type=button title="Copy link" class="btn btn-default share-link-btn" data-clipboard-text=https://thecodinglove.com/when-my-coworker-asks-me-if-i-can-test-his-code data-toggle=tooltip data-placement=right>
-    <svg
-    class=sharing-btn-icon aria-hidden=true focusable=false data-prefix=fas data-icon=link class="svg-inline--fa fa-link fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M326.612 185.391c59.747 59.809 58.927 155.698.36 214.59-.11.12-.24.25-.36.37l-67.2 67.2c-59.27 59.27-155.699 59.262-214.96 0-59.27-59.26-59.27-155.7 0-214.96l37.106-37.106c9.84-9.84 26.786-3.3 27.294 10.606.648 17.722 3.826 35.527 9.69 52.721 1.986 5.822.567 12.262-3.783 16.612l-13.087 13.087c-28.026 28.026-28.905 73.66-1.155 101.96 28.024 28.579 74.086 28.749 102.325.51l67.2-67.19c28.191-28.191 28.073-73.757 0-101.83-3.701-3.694-7.429-6.564-10.341-8.569a16.037 16.037 0 0 1-6.947-12.606c-.396-10.567 3.348-21.456 11.698-29.806l21.054-21.055c5.521-5.521 14.182-6.199 20.584-1.731a152.482 152.482 0 0 1 20.522 17.197zM467.547 44.449c-59.261-59.262-155.69-59.27-214.96 0l-67.2 67.2c-.12.12-.25.25-.36.37-58.566 58.892-59.387 154.781.36 214.59a152.454 152.454 0 0 0 20.521 17.196c6.402 4.468 15.064 3.789 20.584-1.731l21.054-21.055c8.35-8.35 12.094-19.239 11.698-29.806a16.037 16.037 0 0 0-6.947-12.606c-2.912-2.005-6.64-4.875-10.341-8.569-28.073-28.073-28.191-73.639 0-101.83l67.2-67.19c28.239-28.239 74.3-28.069 102.325.51 27.75 28.3 26.872 73.934-1.155 101.96l-13.087 13.087c-4.35 4.35-5.769 10.79-3.783 16.612 5.864 17.194 9.042 34.999 9.69 52.721.509 13.906 17.454 20.446 27.294 10.606l37.106-37.106c59.271-59.259 59.271-155.699.001-214.959z"></path></svg>
-    </button></div></article><article
-    class="blog-post index-blog-post "><h1 class="index-blog-post-title">
-    <a
-    href=https://thecodinglove.com/when-my-code-goes-into-testing-phase>When my code goes into testing phase</a></h1><div
-    class=post-meta-info>
-    <svg
-    class=octicon-svg xmlns=http://www.w3.org/2000/svg width=14 height=16 viewBox="0 0 14 16"><path
-    fill-rule=evenodd d="M10.86 7c-.45-1.72-2-3-3.86-3-1.86 0-3.41 1.28-3.86 3H0v2h3.14c.45 1.72 2 3 3.86 3 1.86 0 3.41-1.28 3.86-3H14V7h-3.14zM7 10.2c-1.22 0-2.2-.98-2.2-2.2 0-1.22.98-2.2 2.2-2.2 1.22 0 2.2.98 2.2 2.2 0 1.22-.98 2.2-2.2 2.2z"/></svg>
-    <span
-    class=post-author>The Coding Love</span> commited on 13 Oct 2022</div><div
-    class=blog-post-content><p><img data-src=https://thecodinglove.com/content/047/CWRslSm.gif></img></p></div><div
-    class=social-share><a
-    href="https://www.facebook.com/sharer/sharer.php?u=https://thecodinglove.com/when-my-code-goes-into-testing-phase" class="js-social-share social-btn" target=_blank>
-    <button
-    id=facebookSharingBtn type=button class="btn btn-default share-fb-btn">
-    <svg
-    class=sharing-btn-icon aria-hidden=true focusable=false data-prefix=fab data-icon=facebook class="svg-inline--fa fa-facebook fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M504 256C504 119 393 8 256 8S8 119 8 256c0 123.78 90.69 226.38 209.25 245V327.69h-63V256h63v-54.64c0-62.15 37-96.48 93.67-96.48 27.14 0 55.52 4.84 55.52 4.84v61h-31.28c-30.8 0-40.41 19.12-40.41 38.73V256h68.78l-11 71.69h-57.78V501C413.31 482.38 504 379.78 504 256z"></path></svg>
-    Facebook
-    </button>
-    </a>
-    <a
-    href="https://twitter.com/intent/tweet?text=When my code goes into testing phase&url=https://thecodinglove.com/when-my-code-goes-into-testing-phase&related=thecodinglove&hashtags=thecodinglove" class=social-btn target=_blank>
-    <button
-    id=twitterSharingBtn type=button class="btn btn-default share-twitter-btn">
-    <svg
-    class=sharing-btn-icon aria-hidden=true focusable=false data-prefix=fab data-icon=twitter class="svg-inline--fa fa-twitter fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M459.37 151.716c.325 4.548.325 9.097.325 13.645 0 138.72-105.583 298.558-298.558 298.558-59.452 0-114.68-17.219-161.137-47.106 8.447.974 16.568 1.299 25.34 1.299 49.055 0 94.213-16.568 130.274-44.832-46.132-.975-84.792-31.188-98.112-72.772 6.498.974 12.995 1.624 19.818 1.624 9.421 0 18.843-1.3 27.614-3.573-48.081-9.747-84.143-51.98-84.143-102.985v-1.299c13.969 7.797 30.214 12.67 47.431 13.319-28.264-18.843-46.781-51.005-46.781-87.391 0-19.492 5.197-37.36 14.294-52.954 51.655 63.675 129.3 105.258 216.365 109.807-1.624-7.797-2.599-15.918-2.599-24.04 0-57.828 46.782-104.934 104.934-104.934 30.213 0 57.502 12.67 76.67 33.137 23.715-4.548 46.456-13.32 66.599-25.34-7.798 24.366-24.366 44.833-46.132 57.827 21.117-2.273 41.584-8.122 60.426-16.243-14.292 20.791-32.161 39.308-52.628 54.253z"></path></svg>
-    Twitter
-    </button>
-    </a><a
-    href="mailto:?body=https://thecodinglove.com/when-my-code-goes-into-testing-phase&subject=" class=social-btn>
-    <button
-    id=emailSharingBtn type=button class="btn btn-default share-email-btn">
-    <svg
-    class=sharing-btn-icon aria-hidden=true focusable=false data-prefix=fas data-icon=envelope class="svg-inline--fa fa-envelope fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M502.3 190.8c3.9-3.1 9.7-.2 9.7 4.7V400c0 26.5-21.5 48-48 48H48c-26.5 0-48-21.5-48-48V195.6c0-5 5.7-7.8 9.7-4.7 22.4 17.4 52.1 39.5 154.1 113.6 21.1 15.4 56.7 47.8 92.2 47.6 35.7.3 72-32.8 92.3-47.6 102-74.1 131.6-96.3 154-113.7zM256 320c23.2.4 56.6-29.2 73.4-41.4 132.7-96.3 142.8-104.7 173.4-128.7 5.8-4.5 9.2-11.5 9.2-18.9v-19c0-26.5-21.5-48-48-48H48C21.5 64 0 85.5 0 112v19c0 7.4 3.4 14.3 9.2 18.9 30.6 23.9 40.7 32.4 173.4 128.7 16.8 12.2 50.2 41.8 73.4 41.4z"></path></svg>
-    Email
-    </button>
-    </a>
-    <button
-    id=copyLinkBtn type=button title="Copy link" class="btn btn-default share-link-btn" data-clipboard-text=https://thecodinglove.com/when-my-code-goes-into-testing-phase data-toggle=tooltip data-placement=right>
-    <svg
-    class=sharing-btn-icon aria-hidden=true focusable=false data-prefix=fas data-icon=link class="svg-inline--fa fa-link fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M326.612 185.391c59.747 59.809 58.927 155.698.36 214.59-.11.12-.24.25-.36.37l-67.2 67.2c-59.27 59.27-155.699 59.262-214.96 0-59.27-59.26-59.27-155.7 0-214.96l37.106-37.106c9.84-9.84 26.786-3.3 27.294 10.606.648 17.722 3.826 35.527 9.69 52.721 1.986 5.822.567 12.262-3.783 16.612l-13.087 13.087c-28.026 28.026-28.905 73.66-1.155 101.96 28.024 28.579 74.086 28.749 102.325.51l67.2-67.19c28.191-28.191 28.073-73.757 0-101.83-3.701-3.694-7.429-6.564-10.341-8.569a16.037 16.037 0 0 1-6.947-12.606c-.396-10.567 3.348-21.456 11.698-29.806l21.054-21.055c5.521-5.521 14.182-6.199 20.584-1.731a152.482 152.482 0 0 1 20.522 17.197zM467.547 44.449c-59.261-59.262-155.69-59.27-214.96 0l-67.2 67.2c-.12.12-.25.25-.36.37-58.566 58.892-59.387 154.781.36 214.59a152.454 152.454 0 0 0 20.521 17.196c6.402 4.468 15.064 3.789 20.584-1.731l21.054-21.055c8.35-8.35 12.094-19.239 11.698-29.806a16.037 16.037 0 0 0-6.947-12.606c-2.912-2.005-6.64-4.875-10.341-8.569-28.073-28.073-28.191-73.639 0-101.83l67.2-67.19c28.239-28.239 74.3-28.069 102.325.51 27.75 28.3 26.872 73.934-1.155 101.96l-13.087 13.087c-4.35 4.35-5.769 10.79-3.783 16.612 5.864 17.194 9.042 34.999 9.69 52.721.509 13.906 17.454 20.446 27.294 10.606l37.106-37.106c59.271-59.259 59.271-155.699.001-214.959z"></path></svg>
-    </button></div></article><article
-    class="blog-post index-blog-post "><h1 class="index-blog-post-title">
-    <a
-    href=https://thecodinglove.com/when-qa-starts-testing-my-code>When QA starts testing my code</a></h1><div
-    class=post-meta-info>
-    <svg
-    class=octicon-svg xmlns=http://www.w3.org/2000/svg width=14 height=16 viewBox="0 0 14 16"><path
-    fill-rule=evenodd d="M10.86 7c-.45-1.72-2-3-3.86-3-1.86 0-3.41 1.28-3.86 3H0v2h3.14c.45 1.72 2 3 3.86 3 1.86 0 3.41-1.28 3.86-3H14V7h-3.14zM7 10.2c-1.22 0-2.2-.98-2.2-2.2 0-1.22.98-2.2 2.2-2.2 1.22 0 2.2.98 2.2 2.2 0 1.22-.98 2.2-2.2 2.2z"/></svg>
-    <span
-    class=post-author>The Coding Love</span> commited on 7 Oct 2022</div><div
-    class=blog-post-content><p><img src=https://thecodinglove.com/content/047/sh8qjSF.gif></img></p></div><div
-    class=social-share><a
-    href="https://www.facebook.com/sharer/sharer.php?u=https://thecodinglove.com/when-qa-starts-testing-my-code" class="js-social-share social-btn" target=_blank>
-    <button
-    id=facebookSharingBtn type=button class="btn btn-default share-fb-btn">
-    <svg
-    class=sharing-btn-icon aria-hidden=true focusable=false data-prefix=fab data-icon=facebook class="svg-inline--fa fa-facebook fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M504 256C504 119 393 8 256 8S8 119 8 256c0 123.78 90.69 226.38 209.25 245V327.69h-63V256h63v-54.64c0-62.15 37-96.48 93.67-96.48 27.14 0 55.52 4.84 55.52 4.84v61h-31.28c-30.8 0-40.41 19.12-40.41 38.73V256h68.78l-11 71.69h-57.78V501C413.31 482.38 504 379.78 504 256z"></path></svg>
-    Facebook
-    </button>
-    </a>
-    <a
-    href="https://twitter.com/intent/tweet?text=When QA starts testing my code&url=https://thecodinglove.com/when-qa-starts-testing-my-code&related=thecodinglove&hashtags=thecodinglove" class=social-btn target=_blank>
-    <button
-    id=twitterSharingBtn type=button class="btn btn-default share-twitter-btn">
-    <svg
-    class=sharing-btn-icon aria-hidden=true focusable=false data-prefix=fab data-icon=twitter class="svg-inline--fa fa-twitter fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M459.37 151.716c.325 4.548.325 9.097.325 13.645 0 138.72-105.583 298.558-298.558 298.558-59.452 0-114.68-17.219-161.137-47.106 8.447.974 16.568 1.299 25.34 1.299 49.055 0 94.213-16.568 130.274-44.832-46.132-.975-84.792-31.188-98.112-72.772 6.498.974 12.995 1.624 19.818 1.624 9.421 0 18.843-1.3 27.614-3.573-48.081-9.747-84.143-51.98-84.143-102.985v-1.299c13.969 7.797 30.214 12.67 47.431 13.319-28.264-18.843-46.781-51.005-46.781-87.391 0-19.492 5.197-37.36 14.294-52.954 51.655 63.675 129.3 105.258 216.365 109.807-1.624-7.797-2.599-15.918-2.599-24.04 0-57.828 46.782-104.934 104.934-104.934 30.213 0 57.502 12.67 76.67 33.137 23.715-4.548 46.456-13.32 66.599-25.34-7.798 24.366-24.366 44.833-46.132 57.827 21.117-2.273 41.584-8.122 60.426-16.243-14.292 20.791-32.161 39.308-52.628 54.253z"></path></svg>
-    Twitter
-    </button>
-    </a><a
-    href="mailto:?body=https://thecodinglove.com/when-qa-starts-testing-my-code&subject=" class=social-btn>
-    <button
-    id=emailSharingBtn type=button class="btn btn-default share-email-btn">
-    <svg
-    class=sharing-btn-icon aria-hidden=true focusable=false data-prefix=fas data-icon=envelope class="svg-inline--fa fa-envelope fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M502.3 190.8c3.9-3.1 9.7-.2 9.7 4.7V400c0 26.5-21.5 48-48 48H48c-26.5 0-48-21.5-48-48V195.6c0-5 5.7-7.8 9.7-4.7 22.4 17.4 52.1 39.5 154.1 113.6 21.1 15.4 56.7 47.8 92.2 47.6 35.7.3 72-32.8 92.3-47.6 102-74.1 131.6-96.3 154-113.7zM256 320c23.2.4 56.6-29.2 73.4-41.4 132.7-96.3 142.8-104.7 173.4-128.7 5.8-4.5 9.2-11.5 9.2-18.9v-19c0-26.5-21.5-48-48-48H48C21.5 64 0 85.5 0 112v19c0 7.4 3.4 14.3 9.2 18.9 30.6 23.9 40.7 32.4 173.4 128.7 16.8 12.2 50.2 41.8 73.4 41.4z"></path></svg>
-    Email
-    </button>
-    </a>
-    <button
-    id=copyLinkBtn type=button title="Copy link" class="btn btn-default share-link-btn" data-clipboard-text=https://thecodinglove.com/when-qa-starts-testing-my-code data-toggle=tooltip data-placement=right>
-    <svg
-    class=sharing-btn-icon aria-hidden=true focusable=false data-prefix=fas data-icon=link class="svg-inline--fa fa-link fa-w-16" role=img xmlns=http://www.w3.org/2000/svg viewBox="0 0 512 512"><path
-    fill=currentColor d="M326.612 185.391c59.747 59.809 58.927 155.698.36 214.59-.11.12-.24.25-.36.37l-67.2 67.2c-59.27 59.27-155.699 59.262-214.96 0-59.27-59.26-59.27-155.7 0-214.96l37.106-37.106c9.84-9.84 26.786-3.3 27.294 10.606.648 17.722 3.826 35.527 9.69 52.721 1.986 5.822.567 12.262-3.783 16.612l-13.087 13.087c-28.026 28.026-28.905 73.66-1.155 101.96 28.024 28.579 74.086 28.749 102.325.51l67.2-67.19c28.191-28.191 28.073-73.757 0-101.83-3.701-3.694-7.429-6.564-10.341-8.569a16.037 16.037 0 0 1-6.947-12.606c-.396-10.567 3.348-21.456 11.698-29.806l21.054-21.055c5.521-5.521 14.182-6.199 20.584-1.731a152.482 152.482 0 0 1 20.522 17.197zM467.547 44.449c-59.261-59.262-155.69-59.27-214.96 0l-67.2 67.2c-.12.12-.25.25-.36.37-58.566 58.892-59.387 154.781.36 214.59a152.454 152.454 0 0 0 20.521 17.196c6.402 4.468 15.064 3.789 20.584-1.731l21.054-21.055c8.35-8.35 12.094-19.239 11.698-29.806a16.037 16.037 0 0 0-6.947-12.606c-2.912-2.005-6.64-4.875-10.341-8.569-28.073-28.073-28.191-73.639 0-101.83l67.2-67.19c28.239-28.239 74.3-28.069 102.325.51 27.75 28.3 26.872 73.934-1.155 101.96l-13.087 13.087c-4.35 4.35-5.769 10.79-3.783 16.612 5.864 17.194 9.042 34.999 9.69 52.721.509 13.906 17.454 20.446 27.294 10.606l37.106-37.106c59.271-59.259 59.271-155.699.001-214.959z"></path></svg>
-    </button></div></article><div
-    class=navigation><div
-    class=float-left></div><div
-    class=float-right>
-    <a
-    href="https://thecodinglove.com/page/2?s=test" class="btn btn-default nav-btn next-posts-btn float-right">Next<i
-    class="fas fa-chevron-right" style=margin-left:7px;></i></a></div></div></div><aside
-    class="col-md-4 sidebar"><div
-    class=sidebar-block>
-    <span
-    style="font-size:11px;color:#636363;font-family: 'IBM Plex Sans', Verdana, Geneva, Tahoma, sans-serif;margin-top:5px;margin-left:5px;width:100%;padding-bottom:10px;">ADVERTISEMENT</span><div
-    style=text-align:center;margin-bottom:5px;>
-    <ins
-    class=adsbygoogle
-    style=display:block
-    data-ad-client=ca-pub-5778613244349789
-    data-ad-slot=9184703807
-    data-ad-format=auto
-    data-full-width-responsive=true></ins> <script>(adsbygoogle = window.adsbygoogle || []).push({});</script> </div></div><div
-    class=sidebar-block><div
-    class="sidebar-header sidebar-header-featured">Featured posts</div><div
-    class="sidebar-block-inset sidebar-block-featured">
-    <a
-    href=https://thecodinglove.com/when-the-intern-leaves-his-first-code-review><div
-    class=sidebar-featured><div
-    class=featured-thumbnail>
-    <img
-    class=lazy src="data:image/svg+xml,%3Csvg%20xmlns='http://www.w3.org/2000/svg'%20viewBox='0%200%20150%20150'%3E%3C/svg%3E" data-src=https://thecodinglove.com/wp-content/uploads/2021/02/Sans-titre-50-150x150.jpg></div><div
-    class=featured-title>
-    When the intern leaves his first code review</div></div>
-    </a>
-    <a
-    href=https://thecodinglove.com/when-i-see-the-error-messages-displaying-right-after-i-shipped-to-production><div
-    class=sidebar-featured><div
-    class=featured-thumbnail>
-    <img
-    class=lazy src="data:image/svg+xml,%3Csvg%20xmlns='http://www.w3.org/2000/svg'%20viewBox='0%200%20150%20150'%3E%3C/svg%3E" data-src=https://thecodinglove.com/wp-content/uploads/2021/02/Sans-titre-44-150x150.jpg></div><div
-    class=featured-title>
-    When I see the error messages displaying right after I shipped to production</div></div>
-    </a>
-    <a
-    href=https://thecodinglove.com/when-i-enter-in-a-loop-while-debugging-with-the-step-by-step-mode><div
-    class=sidebar-featured><div
-    class=featured-thumbnail>
-    <img
-    class=lazy src="data:image/svg+xml,%3Csvg%20xmlns='http://www.w3.org/2000/svg'%20viewBox='0%200%20150%20150'%3E%3C/svg%3E" data-src=https://thecodinglove.com/wp-content/uploads/2021/02/Sans-titre-36-150x150.jpg></div><div
-    class=featured-title>
-    When I enter in a loop while debugging with the step by step mode</div></div>
-    </a>
-    <a
-    href=https://thecodinglove.com/favorite-code-editor-vi-visual-studio-code><div
-    class=sidebar-featured><div
-    class=featured-thumbnail>
-    <img
-    class=lazy src="data:image/svg+xml,%3Csvg%20xmlns='http://www.w3.org/2000/svg'%20viewBox='0%200%20150%20150'%3E%3C/svg%3E" data-src=https://thecodinglove.com/wp-content/uploads/2021/02/Sans-titre-26-150x150.jpg></div><div
-    class=featured-title>
-    When someone asks me what my favorite code editor is</div></div>
-    </a></div></div><div
-    class=sidebar-block><div
-    class="sidebar-header sidebar-header-facebook">Facebook Page</div><div
-    class=sidebar-block-inset><p>Add programmer humor to your Facebook feed by liking the page!</p>
-    <a
-    href=https://facebook.com/thecodinglove target=_blank>
-    <button
-    id=facebookSubscribeBtn type=button class="btn btn-default sidebar-btn sidebar-btn-facebook">
-    Like the page
-    </button>
-    </a></div></div><div
-    class=sidebar-block><div
-    class="sidebar-header sidebar-header-twitter">Twitter</div><div
-    class=sidebar-block-inset><p>Join The Coding Love's community on Twitter</p>
-    <a
-    href="https://twitter.com/intent/follow?screen_name=thecodinglove" target=_blank>
-    <button
-    id=twitterFollowBtn type=button class="btn btn-default sidebar-btn sidebar-btn-twitter">
-    Follow @thecodinglove
-    </button>
-    </a></div></div></aside></div></main><div
-    class=blog-footer><div
-    class=footer-links><p>
-    <a
-    class=footer-links-spaced href=/contact>Contact</a>
-    <a
-    class=footer-links-spaced href=/terms>Terms</a>
-    <a
-    href=/privacy>Privacy</a></p><p
-    class=footer-credits>
-    <span
-    style=font-weight:500>The Coding Love</span>
-    &copy; 2022 Nicolas Lecointre, all rights reserved.</p></div></div> <script>document.addEventListener( 'wpcf7mailsent', function( event ) {
-            if( "fb_pxl_code" in event.detail.apiResponse){
-              eval(event.detail.apiResponse.fb_pxl_code);
-            }
-          }, false );</script> <div
-    id=fb-pxl-ajax-code></div> <script src=https://stats.wp.com/e-202247.js defer></script> <script>_stq = window._stq || [];
-    		_stq.push([ 'view', {v:'ext',blog:'146794450',post:'0',tz:'1',srv:'thecodinglove.com',j:'1:11.5.1'} ]);
-    		_stq.push([ 'clickTrackerInit', '146794450', '0' ]);</script> <link
-    rel=stylesheet href=https://thecodinglove.com/wp-content/cache/minify/5bb47.css media=all> <script src=https://thecodinglove.com/wp-content/cache/minify/ff88c.js></script> <script>function fire_cookie() {
-        window.cookieconsent.initialise({
-            "palette": {
-                "popup": {
-                    "background": "#007acc",
-                    "text": "#fff"
-    			},
-                "button": {
-                    "background": "#fff",
-                    "text": "#007acc"
-                }
-            },
-            "content": {
-                "message": "This website uses cookies to ensure you get the best experience on our website.",
-                "dismiss": "<i>Got it!</i>",
-                "link": "Learn more",
-                "href": "https://thecodinglove.com/privacy"
-            }
-        })
-    }
-    window.onload = fire_cookie;
-
-    window.addEventListener("load", function(){
-        ${'$'}('#submitAnotherModal').on('hide.bs.modal', function (e) { ${'$'}('#tcl-submit-form')[0].reset(); });
-        ${'$'}('#backHomeModal').on('hide.bs.modal', function (e) { ${'$'}('#tcl-submit-form')[0].reset(); });
-    });</script> <script>var selectedModal;
-    function showModal() {
-        var modal_id;
-        
-        if (selectedModal == 0) {
-            modal_id = '#submitAnotherModal';
-        } else {
-            modal_id = '#backHomeModal';
-        }
-        
-        ${'$'}(modal_id).modal('show');
-    }
-    new Clipboard('.share-link-btn');
-
-    function windowPopup(url, width, height) {
-        var left = (screen.width / 2) - (width / 2),
-        top = (screen.height / 2) - (height / 2);
-        
-        window.open(
-            url,
-            "",
-            "menubar=no,toolbar=no,resizable=yes,scrollbars=yes,width=" + width + ",height=" + height + ",top=" + top + ",left=" + left
-        );
-    }
-
-    ${'$'}(".js-social-share").on("click", function(e) {
-        e.preventDefault();
-        windowPopup(${'$'}(this).attr("href"), 500, 300);
-    });
-    console.log('Welcome on the_coding_love();');</script> <script>window.w3tc_lazyload=1,window.lazyLoadOptions={elements_selector:".lazy",callback_loaded:function(t){var e;try{e=new CustomEvent("w3tc_lazyload_loaded",{detail:{e:t}})}catch(a){(e=document.createEvent("CustomEvent")).initCustomEvent("w3tc_lazyload_loaded",!1,!1,{e:t})}window.dispatchEvent(e)}}</script><script async src=https://thecodinglove.com/wp-content/cache/minify/1615d.js></script> </body></html>
-""".trimIndent()
