+++ kmp/kmp-common-firebase/build.gradle.kts
@@ -1,4 +1,3 @@
-
 import com.gchristov.thecodinglove.gradleplugins.Deps
 import java.io.FileInputStream
 import java.util.*
