--- kmp/kmp-common-network/src/commonMain/kotlin/com/gchristov/thecodinglove/kmpcommonfirebase/CommonNetworkModule.kt
@@ -1,37 +1,0 @@
-package com.gchristov.thecodinglove.kmpcommonfirebase
-
-import com.gchristov.thecodinglove.kmpcommondi.DiModule
-import com.gchristov.thecodinglove.kmpcommondi.inject
-import io.ktor.client.*
-import io.ktor.client.plugins.contentnegotiation.*
-import io.ktor.client.plugins.logging.*
-import io.ktor.serialization.kotlinx.json.*
-import kotlinx.serialization.json.Json
-import org.kodein.di.DI
-import org.kodein.di.bindSingleton
-
-object CommonNetworkModule : DiModule() {
-    override fun name() = "kmp-common-network"
-
-    override fun bindLocalDependencies(builder: DI.Builder) {
-        builder.apply {
-            bindSingleton { provideHttpClient() }
-        }
-    }
-
-    fun injectHttpClient(): HttpClient = inject()
-
-    private fun provideHttpClient() = HttpClient {
-        install(ContentNegotiation) {
-            json(
-                Json {
-                    ignoreUnknownKeys = true
-                }
-            )
-        }
-        install(Logging) {
-            logger = Logger.SIMPLE
-            level = LogLevel.ALL
-        }
-    }
-}
