+++ kmp/kmp-htmlparse-data/src/commonMain/kotlin/com/gchristov/thecodinglove/kmphtmlparsedata/HtmlPost.kt
@@ -1,0 +1,7 @@
+package com.gchristov.thecodinglove.kmphtmlparsedata
+
+data class HtmlPost(
+    val title: String,
+    val url: String,
+    val imageUrl: String,
+)
