+++ kmp/kmp-htmlparse-testfixtures/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    id("kmp-module-plugin")
+}
