+++ kmp/kmp-htmlparse/build.gradle.kts
@@ -1,0 +1,18 @@
+plugins {
+    id("kmp-module-plugin")
+}
+
+kotlin {
+    sourceSets {
+        val commonMain by getting {
+            dependencies {
+                api(projects.kmpHtmlparseData)
+            }
+        }
+        val commonTest by getting {
+            dependencies {
+                implementation(projects.kmpHtmlparseTestfixtures)
+            }
+        }
+    }
+}
