+++ kmp/kmp-htmlparse/src/commonMain/kotlin/com/gchristov/thecodinglove/kmphtmlparse/HtmlPostParser.kt
@@ -1,0 +1,9 @@
+package com.gchristov.thecodinglove.kmphtmlparse
+
+import com.gchristov.thecodinglove.kmphtmlparsedata.HtmlPost
+
+interface HtmlPostParser {
+    suspend fun parseTotalPosts(content: String): Int
+
+    suspend fun parsePosts(content: String): List<HtmlPost>
+}
