+++ kmp/kmp-htmlparse/src/jsMain/kotlin/com/gchristov/thecodinglove/kmphtmlparse/JavascriptSearchDataModule.kt
@@ -1,0 +1,7 @@
+package com.gchristov.thecodinglove.kmphtmlparse
+
+import kotlinx.coroutines.Dispatchers
+
+internal actual fun provideHtmlPostParser(): HtmlPostParser = RealHtmlPostParser(
+    dispatcher = Dispatchers.Default
+)
