--- kmp/kmp-search-data/src/commonMain/kotlin/com/gchristov/thecodinglove/kmpsearchdata/PostParser.kt
@@ -1,7 +1,0 @@
-package com.gchristov.thecodinglove.kmpsearchdata
-
-interface PostParser {
-    suspend fun parseTotalPosts(content: String): Int
-
-    suspend fun parsePosts(content: String): List<Post>
-}
