+++ kmp/kmp-search-data/src/commonMain/kotlin/com/gchristov/thecodinglove/kmpsearchdata/SearchDataModule.kt
@@ -1,20 +1,43 @@
 package com.gchristov.thecodinglove.kmpsearchdata
 
 import com.gchristov.thecodinglove.kmpcommondi.DiModule
-import com.gchristov.thecodinglove.kmpcommondi.inject
+import com.gchristov.thecodinglove.kmpcommonnetwork.CommonNetworkModule
+import com.gchristov.thecodinglove.kmphtmlparse.HtmlParseModule
+import com.gchristov.thecodinglove.kmphtmlparse.HtmlPostParser
+import io.ktor.client.*
 import org.kodein.di.DI
-import org.kodein.di.bindProvider
+import org.kodein.di.bindSingleton
+import org.kodein.di.instance
 
 object SearchDataModule : DiModule() {
     override fun name() = "kmp-search-data"
 
     override fun bindLocalDependencies(builder: DI.Builder) {
         builder.apply {
-            bindProvider { providePostParser() }
+            bindSingleton { provideSearchApi(client = instance()) }
+            bindSingleton {
+                provideSearchRepository(
+                    api = instance(),
+                    htmlPostParser = instance()
+                )
+            }
         }
     }
 
-    fun injectPostParser(): PostParser = inject()
-}
+    override fun moduleDependencies(): List<DI.Module> {
+        return listOf(
+            CommonNetworkModule.module,
+            HtmlParseModule.module
+        )
+    }
 
-internal expect fun providePostParser(): PostParser
+    private fun provideSearchApi(client: HttpClient) = SearchApi(client)
+
+    private fun provideSearchRepository(
+        api: SearchApi,
+        htmlPostParser: HtmlPostParser
+    ): SearchRepository = RealSearchRepository(
+        apiService = api,
+        htmlPostParser = htmlPostParser
+    )
+}
