+++ kmp/kmp-search-data/src/commonMain/kotlin/com/gchristov/thecodinglove/kmpsearchdata/SearchRepository.kt
@@ -1,0 +1,10 @@
+package com.gchristov.thecodinglove.kmpsearchdata
+
+interface SearchRepository {
+    suspend fun getTotalPosts(query: String): Int
+
+    suspend fun search(
+        page: Int,
+        query: String
+    ): List<Post>
+}
