--- kmp/kmp-search-data/src/jsMain/kotlin/com/gchristov/thecodinglove/kmpsearchdata/JavascriptSearchDataModule.kt
@@ -1,6 +1,0 @@
-package com.gchristov.thecodinglove.kmpsearchdata
-
-import kotlinx.coroutines.Dispatchers
-
-internal actual fun providePostParser(): PostParser =
-    HtmlPostParser(dispatcher = Dispatchers.Default)
