+++ kmp/kmp-search-testfixtures/build.gradle.kts
@@ -1,3 +1,13 @@
 plugins {
     id("kmp-module-plugin")
 }
+
+kotlin {
+    sourceSets {
+        val commonMain by getting {
+            dependencies {
+                implementation(projects.kmpSearchData)
+            }
+        }
+    }
+}
