+++ modules.gradle.kts
@@ -8,6 +8,10 @@ val projects = listOf(
     "kmp-common-kotlin",
     "kmp-common-network",
     "kmp-common-test",
+    "kmp-htmlparse",
+    "kmp-htmlparse-data",
+    "kmp-htmlparse-testfixtures",
+    "kmp-search",
     "kmp-search-data",
     "kmp-search-testfixtures",
     "appApiJs",
